This seems to clear up all the problems calculating the canvasItem locations.  They are now tied, by runtime calculations, to the textEdit.

+++ Panels/SourcePanel/SourcePanel.hxx
@@ -99,7 +99,7 @@ public:
   int lastScrollBarValue;
 
   //! The last line height
-  float lastLineHeight;
+  int lastLineHeight;
   
   //! The last calculated visible lines...
   int lastVisibleLines;
